moved col to journalist_app/col.py
